Move dmx output struct to separate module
